backtester/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.1
h11==0.13.0
idna==3.3
numpy==1.22.3
pandas==1.4.2
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
